requirements.txt
# Patent + Innovation Radar: Production Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23

# Database & Caching
psycopg2-binary==2.9.9
redis==5.0.1
qdrant-client==1.16.2
neo4j==5.13.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4

# ML & Embeddings
sentence-transformers==2.2.2
torch==2.1.1
bertopic==0.14.0
lightgbm==4.6.0
transformers==4.35.2

# LLM & Agents
langgraph==0.0.49
anthropic==0.7.8

# Data Quality & Validation
great-expectations==0.18.8
pydantic[email]==2.5.0

# Orchestration & Scheduling
celery==5.3.4
kfp==2.4.0  # Kubeflow Pipelines

# MLOps
mlflow==2.10.2
evidently==0.4.22

# Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
python-json-logger==2.0.7

# API & Web
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# UI
streamlit==1.28.1
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
backoff==2.2.1
python-dateutil==2.8.2
click==8.1.7
pyyaml==6.0.1
tqdm==4.66.1